requirements.txt
google-generativeai
scikit-learn
pandas
python-dotenv
requests
